Wrote write_circuit() method for ProductAnsatz class.
Added _params.py file for class for working with parameters.

Added _test_base_ansatz.py with new unit tests for the BaseAnsatz class.

Overwrote __sum__ method for BaseAnsatz so that two ansatze with the same numbers
of qubits can be added together to return a new ansatz with the concatenation
of the circuits.