New methods and tests for BaseAnsatz.
Changes to BaseAnsatz:
(1) Added a compile method which returns a compiled circuit for a given computer.
(2) Fixed the depth method to return the length of the instructions only (not PRAGMAs or DEFGATEs, etc.)
(3) Added an "add_at" method which applies an input gate at a given set of qubit indices.

New tests for compiling, computing depths, and add_at.