added more _cdata test and altered _cdata slightly
Only change to _cdata was to remove assertion that len(data.shape) >= 1. This
is because data with only a single feature is valid, and it is more conenient
to enter it as np.array([1, 2, 3]) than np.array([[1], [2], [3]]).

New tests ensure that higher dimensional data has right shape, that labeling
works when passing labels or a function, and that data splitting works.