added ascii_circuit creator from pyquil program
Allows you to create a simple ascii circuit representation of your
pyquil program. At this point, it calls the circuits by their pyquil-given
names, but there is a (not fully implemented) rebind option to change this.
Further, rather than having two lines for each qubit that connects the 2 qubit
gates, I've added a symbol ' to denote an asymmetry in asymmetric 2 qubit
gates.