added alt circ drawer that 'stacks' 1 qubit gates
Here, I ran through the program while storing two dicionaries:
gateorder: keeps track of order in which gates appear for each qubit wire
gatenum: keeps a numbering on each qubit to know where to place it

The idea is that 2 qubit gates will take precedent over 1 qubit gates. Hence,
the lack of 1 qubit gates for a wire that later has a 2 qubit gate must be
filled with 'left' padding until it lines up with 2 qubit gate of other wire.
Because of this, I had to inlcude a variable padlength that now depends
on the length of the longest gate name.

For this reason, the next change is to come up with a set of good renames
for gates to shrink down the length of the names.

Further, I need to add support for the POST-compiled programs if that doesn't
already work.