fixed qubit state labels, added rebinding, started on asym
Made qubit kets actually have their state, namely, |0>. Rebinding allows for
things like {'CNOT': '@'} to turn CNOTS into a different symbol.

PROBLEM: Now, the problem with adding asym at this stage is that a dictionary
seems to destroy the information about which qubit is listed first in the gate.
To keep track of this information, I will need to add the asymmetry in the
checking of the gate.out() command wherein I can check the internally passed
parameters such as the qubits it is acting on. I think it can be done, but I'm
about to get on my plane...